+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 - Fix unit test failures caused by Tokenize performance optimization code
+- Fix Rumor bindings no longer attempting to convert arguments
 
 ## [1.0.1] - 2017-12-19